+++ CHANGELOG.md
@@ -18,6 +18,9 @@ for further information about branching and tagging conventions.
 - Displaying asset policies & asset type on the `AssetDetails` page
 - `record` getter to the `AssetRecord`
 - Added confirmation to cancel token creation request
+- CLI config passing
+- Added module enabling/disabling
+- Added module schemes
 
 ### Changed
 - Now using new account endpoint for loading the fees